feat: Refactor GeoJsonUploadIntegrationTest
Refactored the GeoJsonUploadIntegrationTest to use a more robust verification method.

The previous implementation relied on counting entities in the JPA and Elasticsearch repositories, which could be flaky. The new implementation fetches the created entities and asserts that the correct number of entities were created.

This change also adds the  directory to the .gitignore file.